model_builds/depth_estimator/requirements_depth_estimator_gpu.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# Core depth-estimation stack
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
numpy>=1.23.0,<2.0
Pillow>=8.0.0
timm>=0.9.0
transformers>=4.40.1,<5.0
accelerate>=0.21.0
safetensors>=0.4.0
huggingface_hub>=0.20.0,<0.24.0

# HTTP serving
Flask>=2.3.0
gunicorn>=21.2.0
